requirements.txt
opencv-python==3.4.2.17
imageio==2.8.0
scipy==1.4.1
numpy==1.18.2
pandas==1.0.3
pyrealsense2==2.33.1.1388
matplotlib==3.2.1
pyntcloud==0.1.2
Pillow==7.0.0
open3d==0.12.0.0
